Developing findAll Brasileirao Class

+++ Backend/campeonato/src/main/java/br/com/brasileirao/campeonato/repository/BrasileiraoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrasileiraoRepository extends JpaRepository<Brasileirao, Integer> {
+public interface BrasileiraoRepository extends JpaRepository<Brasileirao, Long> {
 }
